Ensure that Q7 implementations truncate fractional seconds
Q7 needs to report the minutes passed -- this should be done without any
rounding, i.e. 60000 milliseconds must pass, 59999 milliseconds are
insufficient.

By default, Postgres rounds a number when casting a float to an integer.
This commit add 'floor' to avoid to ensure the number is truncated.

The Neo4j/Cypher implementation was changed: it previously used an
explicit 'round' function and now uses 'floor'.

The investigation of this problem led to the following bug report:
https://github.com/duckdb/duckdb/issues/3497