Bump Neo4j version, rework Q14 implementation's GDS calls
The Graph Data Science API deprecated the use of anonymous graph
projection, hence, we need to project a named graph for Q14.
I implemented this with a hack using 3 "sub"-queries by chaining
together the procedure calls using the 'WITH count(*) AS dummy' trick.
This may cause an issue when run concurrently but works fine for
validation.